Refactor: 업적 부분을 achievements 모듈로 분리
