+++ src/achievements/achievements.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from '@nestjs/common';
+
+@Controller('achievements')
+export class AchievementsController {}
