+++ src/achievements/entities/user-achievement.entity.ts
@@ -1,6 +1,6 @@
+import { User } from 'src/users/entities/user.entity';
 import { CreateDateColumn, Entity, ManyToOne } from 'typeorm';
 import { Achievement } from './achievement.entity';
-import { User } from './user.entity';
 
 @Entity()
 export class UserAchievement {
