+++ src/channels/channels.module.ts
@@ -8,8 +8,6 @@ import { UsersRepository } from 'src/users/repositories/users.repository';
 import { UsersService } from 'src/users/users.service';
 import { ChatsRepository } from './repositories/chats.repository';
 import { EventsModule } from 'src/events/events.module';
-import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
-import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
 
 @Module({
   imports: [
@@ -17,8 +15,6 @@ import { AchievementsRepository } from 'src/users/repositories/achievement.repos
       ChannelsRepository,
       MembershipsRepository,
       UsersRepository,
-      AchievementsRepository,
-      UserAchievementsRepository,
       ChatsRepository,
     ]),
     EventsModule,
