+++ src/dms/dms.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { AchievementsModule } from 'src/achievements/achievements.module';
 import { EventsModule } from 'src/events/events.module';
 import { FriendshipsRepository } from 'src/friendships/friendships.repository';
 import { FriendshipsService } from 'src/friendships/friendships.service';
-import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
-import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
 import { UsersRepository } from 'src/users/repositories/users.repository';
 import { UsersService } from 'src/users/users.service';
 import { DmsController } from './dms.controller';
@@ -16,11 +15,10 @@ import { DmsService } from './dms.service';
     TypeOrmModule.forFeature([
       DmsRepository,
       UsersRepository,
-      AchievementsRepository,
-      UserAchievementsRepository,
       FriendshipsRepository,
     ]),
     EventsModule,
+    AchievementsModule,
   ],
   controllers: [DmsController],
   providers: [DmsService, UsersService, FriendshipsService],
