+++ src/friendships/friendships.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AchievementsRepository } from 'src/users/repositories/achievement.repository';
-import { UserAchievementsRepository } from 'src/users/repositories/user-achievement.repository';
+import { AchievementsModule } from 'src/achievements/achievements.module';
 import { UsersRepository } from 'src/users/repositories/users.repository';
 import { UsersService } from 'src/users/users.service';
 import { FriendshipsController } from './friendships.controller';
@@ -10,12 +9,8 @@ import { FriendshipsService } from './friendships.service';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([
-      FriendshipsRepository,
-      UsersRepository,
-      AchievementsRepository,
-      UserAchievementsRepository,
-    ]),
+    TypeOrmModule.forFeature([FriendshipsRepository, UsersRepository]),
+    AchievementsModule,
   ],
   controllers: [FriendshipsController],
   providers: [FriendshipsService, UsersService],
